feat(sync): Extract code of attachment syncing from main process.
